requirements_test.txt
black==22.1.0
pylint==2.12.2
pylint-strict-informational==0.1
pytest==7.0.1
python-dotenv==1.0.0
